refactor : role and permission cleanup code

+++ src/modules/auth/components/permission/use-case/delete-permission.use-case.ts
@@ -7,15 +7,20 @@ import {
   DeletePermissionInput,
   DeletePermissionOutput,
 } from '@/modules/auth/components/permission/dto/delete-permission.dto';
+import { PermissionHelepr } from '@/modules/auth/components/permission/helper/permission-helper';
 
 @Injectable()
 export class DeletePermissionUseCase {
-  constructor(private readonly commandBus: CommandBus) {}
+  constructor(
+    private readonly commandBus: CommandBus,
+    private readonly permissionHelepr: PermissionHelepr,
+  ) {}
 
   async deletePermission(
     input: DeletePermissionInput,
   ): Promise<DeletePermissionOutput> {
     try {
+      await this.permissionHelepr.validatePermissionId(input.permissionId);
       await this.commandBus.execute(new DeletePermissionCommand(input));
       return { success: true };
     } catch (err) {
